+++ 1.21.5/common/src/main/java/net/xstopho/resource_backpacks/mixin/client/LivingEntityRendererMixin.java
@@ -1,13 +1,11 @@
 package net.xstopho.resource_backpacks.mixin.client;
 
-import net.minecraft.client.model.EntityModel;
-import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.client.renderer.entity.layers.RenderLayer;
 import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
 import net.xstopho.resource_backpacks.backpack.api.BackpackHolder;
-import net.xstopho.resource_backpacks.client.BackpackRenderLayer;
 import net.xstopho.resource_backpacks.client.util.BackpackRenderState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
@@ -21,17 +19,12 @@ public abstract class LivingEntityRendererMixin {
     @Invoker("addLayer")
     public abstract boolean invokeAddLayer(RenderLayer<?, ?> feature);
 
-    @Inject(at = @At("RETURN"), method = "<init>")
-    public void init(EntityRendererProvider.Context context, EntityModel<?> model, float shadowRadius, CallbackInfo info) {
-        this.invokeAddLayer(new BackpackRenderLayer<>((LivingEntityRenderer) (Object) this, context.getModelSet()));
-    }
 
     @Inject(method = "extractRenderState", at = @At("TAIL"))
     public void resource_backpacks$extractRenderState(LivingEntity livingEntity, LivingEntityRenderState renderState, float partialTicks, CallbackInfo info) {
-        ((BackpackHolder) livingEntity).getBackpack().ifPresent(itemStack -> {
-            if (renderState instanceof BackpackRenderState backpackRenderState) {
-                backpackRenderState.setBackpack(itemStack);
-            }
-        });
+        ItemStack itemStack = ((BackpackHolder) livingEntity).getBackpack();
+        if (renderState instanceof BackpackRenderState backpackRenderState) {
+            backpackRenderState.setBackpack(itemStack);
+        }
     }
 }