+++ 1.21.5/common/src/main/java/net/xstopho/resource_backpacks/mixin/common/LivingEntityMixin.java
@@ -21,20 +21,20 @@ import java.util.Optional;
 public abstract class LivingEntityMixin extends Entity implements BackpackHolder {
 
     @Unique
-    private Optional<ItemStack> backpack = Optional.empty();
+    private ItemStack backpack = ItemStack.EMPTY;
 
     public LivingEntityMixin(EntityType<?> entityType, Level level) {
         super(entityType, level);
     }
 
     @Override
-    public Optional<ItemStack> getBackpack() {
+    public ItemStack getBackpack() {
         return backpack;
     }
 
     @Override
     public void setBackpack(ItemStack backpack) {
-        this.backpack = Optional.of(backpack);
+        this.backpack = backpack;
     }
 
     @Inject(method = "die", at = @At("TAIL"))