+++ CHANGELOG.md
@@ -3,3 +3,5 @@
   - fixing issue where Backpacks disappear when 
     killing the Ender Dragon the first time
   - fixing render issue when Fresh Moves is installed
+  - adding new DataComponent to fit the new TooltipProvider System
+    - Compatible with older versions, so no Items get lost