+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### v1.1.0
   - update to 1.21.5
   - fixing issue where Backpacks disappear when 
-    killing the Ender Dragon the first time
+    respawning after the credits
   - fixing render issue when Fresh Moves is installed
   - adding new DataComponent to fit the new TooltipProvider System
     - Compatible with older versions, so no Items get lost