+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-### v1.1.1
-  - disable Backpack placement for 1.21.5 due to a Bug that causes Item Loss
+### v1.1.2
+  - fixing Bug where Backpacks drop there Contents when the Player destroy the Backpack Block
+  -